--- poly_commit/src/orion/pcs_impl.rs
@@ -1,174 +1,0 @@
-use std::iter;
-use std::ops::Mul;
-
-use arith::{Field, SimdField};
-use polynomials::{EqPolynomial, MultiLinearPoly};
-use transcript::Transcript;
-
-use crate::{traits::TensorCodeIOPPCS, PCS_SOUNDNESS_BITS};
-
-use super::utils::*;
-
-#[inline(always)]
-pub(crate) fn commit_encoded<F, PackF>(
-    pk: &OrionSRS,
-    packed_evals: &[PackF],
-    scratch_pad: &mut OrionScratchPad<F, PackF>,
-    packed_rows: usize,
-    msg_size: usize,
-) -> OrionResult<OrionCommitment>
-where
-    F: Field,
-    PackF: SimdField<Scalar = F>,
-{
-    // NOTE: packed codeword buffer and encode over packed field
-    let mut packed_interleaved_codewords = vec![PackF::ZERO; packed_rows * pk.codeword_len()];
-    packed_evals
-        .chunks(msg_size)
-        .zip(packed_interleaved_codewords.chunks_mut(pk.codeword_len()))
-        .try_for_each(|(evals, codeword)| pk.code_instance.encode_in_place(evals, codeword))?;
-
-    // NOTE: transpose codeword s.t., the matrix has codewords being columns
-    let mut scratch = vec![PackF::ZERO; packed_rows * pk.codeword_len()];
-    transpose_in_place(&mut packed_interleaved_codewords, &mut scratch, packed_rows);
-    drop(scratch);
-
-    // NOTE: commit the interleaved codeword
-    // we just directly commit to the packed field elements to leaves
-    // Also note, when codeword is not power of 2 length, pad to nearest po2
-    // to commit by merkle tree
-    if !packed_interleaved_codewords.len().is_power_of_two() {
-        let aligned_po2_len = packed_interleaved_codewords.len().next_power_of_two();
-        packed_interleaved_codewords.resize(aligned_po2_len, PackF::ZERO);
-    }
-    scratch_pad.interleaved_alphabet_commitment =
-        tree::Tree::compact_new_with_packed_field_elems::<F, PackF>(packed_interleaved_codewords);
-
-    Ok(scratch_pad.interleaved_alphabet_commitment.root())
-}
-
-#[inline(always)]
-pub(crate) fn orion_mt_openings<F, EvalF, ComPackF, T>(
-    pk: &OrionSRS,
-    transcript: &mut T,
-    scratch_pad: &OrionScratchPad<F, ComPackF>,
-) -> Vec<tree::RangePath>
-where
-    F: Field,
-    EvalF: Field + From<F> + Mul<F, Output = EvalF>,
-    ComPackF: SimdField<Scalar = F>,
-    T: Transcript<EvalF>,
-{
-    let leaves_in_range_opening = OrionSRS::LEAVES_IN_RANGE_OPENING;
-
-    // NOTE: MT opening for point queries
-    let query_num = pk.query_complexity(PCS_SOUNDNESS_BITS);
-    let query_indices = transcript.generate_challenge_index_vector(query_num);
-    query_indices
-        .iter()
-        .map(|qi| {
-            let index = *qi % pk.codeword_len();
-            let left = index * leaves_in_range_opening;
-            let right = left + leaves_in_range_opening - 1;
-
-            scratch_pad
-                .interleaved_alphabet_commitment
-                .range_query(left, right)
-        })
-        .collect()
-}
-
-pub fn orion_verify<F, EvalF, ComPackF, OpenPackF, T>(
-    vk: &OrionSRS,
-    commitment: &OrionCommitment,
-    point: &[EvalF],
-    evaluation: EvalF,
-    transcript: &mut T,
-    proof: &OrionProof<EvalF>,
-) -> bool
-where
-    F: Field,
-    EvalF: Field + From<F> + Mul<F, Output = EvalF>,
-    ComPackF: SimdField<Scalar = F>,
-    OpenPackF: SimdField<Scalar = F>,
-    T: Transcript<EvalF>,
-{
-    let (row_num, msg_size) = OrionSRS::evals_shape::<F>(point.len());
-    let num_of_vars_in_msg = msg_size.ilog2() as usize;
-
-    // NOTE: working on evaluation response, evaluate the rest of the response
-    let mut scratch = vec![EvalF::ZERO; msg_size];
-    let final_eval = MultiLinearPoly::evaluate_with_buffer(
-        &proof.eval_row,
-        &point[..num_of_vars_in_msg],
-        &mut scratch,
-    );
-    if final_eval != evaluation {
-        return false;
-    }
-
-    // NOTE: working on proximity responses, draw random linear combinations
-    // then draw query points from fiat shamir transcripts
-    let proximity_test_num = vk.proximity_repetitions::<EvalF>(PCS_SOUNDNESS_BITS);
-    let random_linear_combinations: Vec<Vec<EvalF>> = (0..proximity_test_num)
-        .map(|_| transcript.generate_challenge_field_elements(row_num))
-        .collect();
-    let query_num = vk.query_complexity(PCS_SOUNDNESS_BITS);
-    let query_indices = transcript.generate_challenge_index_vector(query_num);
-
-    // NOTE: check consistency in MT in the opening trees and against the commitment tree
-    let leaf_range = row_num / tree::leaf_adic::<F>();
-    let mt_consistency =
-        query_indices
-            .iter()
-            .zip(proof.query_openings.iter())
-            .all(|(&qi, range_path)| {
-                let index = qi % vk.codeword_len();
-                range_path.verify(commitment) && index == range_path.left / leaf_range
-            });
-    if !mt_consistency {
-        return false;
-    }
-
-    // NOTE: prepare the interleaved alphabets from the MT paths,
-    // but pack them back into look up table acceptable formats
-    let packed_interleaved_alphabets: Vec<_> = proof
-        .query_openings
-        .iter()
-        .map(|p| -> Vec<_> {
-            p.unpack_field_elems::<F, ComPackF>()
-                .chunks(OpenPackF::PACK_SIZE)
-                .map(OpenPackF::pack)
-                .collect()
-        })
-        .collect();
-
-    // NOTE: encode the proximity/evaluation responses,
-    // check againts all challenged indices by check alphabets against
-    // linear combined interleaved alphabet
-    let mut luts = SubsetSumLUTs::new(OpenPackF::PACK_SIZE, row_num / OpenPackF::PACK_SIZE);
-    assert_eq!(row_num % OpenPackF::PACK_SIZE, 0);
-
-    let eq_linear_combination = EqPolynomial::build_eq_x_r(&point[num_of_vars_in_msg..]);
-    random_linear_combinations
-        .iter()
-        .zip(proof.proximity_rows.iter())
-        .chain(iter::once((&eq_linear_combination, &proof.eval_row)))
-        .all(|(rl, msg)| {
-            let codeword = match vk.code_instance.encode(msg) {
-                Ok(c) => c,
-                _ => return false,
-            };
-
-            luts.build(rl);
-
-            query_indices
-                .iter()
-                .zip(packed_interleaved_alphabets.iter())
-                .all(|(&qi, interleaved_alphabet)| {
-                    let index = qi % vk.codeword_len();
-                    let alphabet = luts.lookup_and_sum(interleaved_alphabet);
-                    alphabet == codeword[index]
-                })
-        })
-}
