+++ poly_commit/src/orion/utils.rs
@@ -1,9 +1,10 @@
-use std::marker::PhantomData;
+use std::{marker::PhantomData, ops::Mul};
 
 use arith::{Field, FieldSerdeError, SimdField};
 use thiserror::Error;
+use transcript::Transcript;
 
-use crate::{traits::TensorCodeIOPPCS, StructuredReferenceString};
+use crate::{traits::TensorCodeIOPPCS, StructuredReferenceString, PCS_SOUNDNESS_BITS};
 
 use super::linear_code::{OrionCode, OrionCodeParameter};
 
@@ -99,6 +100,75 @@ pub struct OrionProof<EvalF: Field> {
     pub eval_row: Vec<EvalF>,
     pub proximity_rows: Vec<Vec<EvalF>>,
     pub query_openings: Vec<tree::RangePath>,
+}
+
+#[inline(always)]
+pub(crate) fn commit_encoded<F, PackF>(
+    pk: &OrionSRS,
+    packed_evals: &[PackF],
+    scratch_pad: &mut OrionScratchPad<F, PackF>,
+    packed_rows: usize,
+    msg_size: usize,
+) -> OrionResult<OrionCommitment>
+where
+    F: Field,
+    PackF: SimdField<Scalar = F>,
+{
+    // NOTE: packed codeword buffer and encode over packed field
+    let mut packed_interleaved_codewords = vec![PackF::ZERO; packed_rows * pk.codeword_len()];
+    packed_evals
+        .chunks(msg_size)
+        .zip(packed_interleaved_codewords.chunks_mut(pk.codeword_len()))
+        .try_for_each(|(evals, codeword)| pk.code_instance.encode_in_place(evals, codeword))?;
+
+    // NOTE: transpose codeword s.t., the matrix has codewords being columns
+    let mut scratch = vec![PackF::ZERO; packed_rows * pk.codeword_len()];
+    transpose_in_place(&mut packed_interleaved_codewords, &mut scratch, packed_rows);
+    drop(scratch);
+
+    // NOTE: commit the interleaved codeword
+    // we just directly commit to the packed field elements to leaves
+    // Also note, when codeword is not power of 2 length, pad to nearest po2
+    // to commit by merkle tree
+    if !packed_interleaved_codewords.len().is_power_of_two() {
+        let aligned_po2_len = packed_interleaved_codewords.len().next_power_of_two();
+        packed_interleaved_codewords.resize(aligned_po2_len, PackF::ZERO);
+    }
+    scratch_pad.interleaved_alphabet_commitment =
+        tree::Tree::compact_new_with_packed_field_elems::<F, PackF>(packed_interleaved_codewords);
+
+    Ok(scratch_pad.interleaved_alphabet_commitment.root())
+}
+
+#[inline(always)]
+pub(crate) fn orion_mt_openings<F, EvalF, ComPackF, T>(
+    pk: &OrionSRS,
+    transcript: &mut T,
+    scratch_pad: &OrionScratchPad<F, ComPackF>,
+) -> Vec<tree::RangePath>
+where
+    F: Field,
+    EvalF: Field + From<F> + Mul<F, Output = EvalF>,
+    ComPackF: SimdField<Scalar = F>,
+    T: Transcript<EvalF>,
+{
+    let leaves_in_range_opening = OrionSRS::LEAVES_IN_RANGE_OPENING;
+
+    // NOTE: MT opening for point queries
+    let query_num = pk.query_complexity(PCS_SOUNDNESS_BITS);
+    let query_indices = transcript.generate_challenge_index_vector(query_num);
+    query_indices
+        .iter()
+        .map(|qi| {
+            let index = *qi % pk.codeword_len();
+            let left = index * leaves_in_range_opening;
+            let right = left + leaves_in_range_opening - 1;
+
+            scratch_pad
+                .interleaved_alphabet_commitment
+                .range_query(left, right)
+        })
+        .collect()
 }
 
 /*
